runとappに役割分割
